+++ agent-core/src/main/java/io/kindling/agent/instrument/aspect/advice/AdviceScaner.java
@@ -19,6 +19,7 @@ package io.kindling.agent.instrument.aspect.advice;
 import java.util.ArrayList;
 import java.util.List;
 
+import io.kindling.agent.api.MethodModifier;
 import io.kindling.agent.deps.org.objectweb.asm.AnnotationVisitor;
 import io.kindling.agent.deps.org.objectweb.asm.ClassVisitor;
 import io.kindling.agent.deps.org.objectweb.asm.Opcodes;
@@ -39,6 +40,7 @@ public class AdviceScaner extends ClassVisitor {
     private String adviceName;
     private boolean isInterface;
     private AnnotatedAdviceInfo adviceInfo;
+    private List<MethodModifier> matchMethodModifiers = new ArrayList<MethodModifier>();
     private List<String> matchClasses = new ArrayList<String>();
     private List<String> matchMethods = new ArrayList<String>();
     private List<String> matchParams = new ArrayList<String>();
@@ -63,6 +65,14 @@ public class AdviceScaner extends ClassVisitor {
                 av = new AnnotationVisitor(api, av) {
                     public AnnotationVisitor visitArray(String name) {
                         AnnotationVisitor av = super.visitArray(name);
+                        if ("matchModifiers".equals(name)) {
+                            av = new AnnotationVisitor(api, av) {
+                                public void visitEnum(String name, String desc, String value) {
+                                    super.visitEnum(name, desc, value);
+                                    matchMethodModifiers.add(MethodModifier.of(value));
+                                }
+                            };
+                        }
                         if ("matchClasses".equals(name)) {
                             av = new AnnotationVisitor(api, av) {
                                 public void visit(String name, Object value) {
@@ -127,13 +137,26 @@ public class AdviceScaner extends ClassVisitor {
             if (matchParams.size() > maxNum) {
                 maxNum = matchParams.size();
             }
+            if (matchMethodModifiers.isEmpty()) {
+                for (int i = 0; i < maxNum; i++) {
+                    matchMethodModifiers.add(MethodModifier.ALL);
+                }
+            } else if (matchMethodModifiers.size() == 1) {
+                for (int i = 1; i < maxNum; i++) {
+                    matchMethodModifiers.add(matchMethodModifiers.get(0));
+                }
+            } else if (matchMethodModifiers.size() != maxNum) {
+                ServiceFactory.LOG.info("[Ignore Advice] {} by argNums with [matchMethodModifiers] must be {}.", adviceName, maxNum);
+                return;
+            }
             
-            adviceInfo.setPointCut(new AnnotationPointCut(matchClasses.get(0), matchMethods.get(0), matchParams.get(0)));
+            adviceInfo.setPointCut(new AnnotationPointCut(matchMethodModifiers.get(0), matchClasses.get(0), matchMethods.get(0), matchParams.get(0)));
             for (int i = 1; i < maxNum; i++) {
+                MethodModifier matchMethodModifier = matchMethodModifiers.size() == 1 ? matchMethodModifiers.get(0) :matchMethodModifiers.get(i);
                 String matchClass = matchClasses.size() == 1 ? matchClasses.get(0) : matchClasses.get(i);
                 String matchMethod = matchMethods.size() == 1 ? matchMethods.get(0) : matchMethods.get(i);
                 String matchParam = matchParams.size() == 1 ? matchParams.get(0) : matchParams.get(i);
-                adviceInfo.getPointCut().addMatcher(matchClass, matchMethod, matchParam);
+                adviceInfo.getPointCut().addMatcher(matchMethodModifier, matchClass, matchMethod, matchParam);
             }
             AdviceContext.addAdviceInfo(adviceInfo);
         }