+++ agent-thirds/third-skywalking/src/main/java/org/apache/skywalking/apm/agent/core/context/trace/StackBasedTracingSpan.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2022 The Kindling Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.skywalking.apm.agent.core.context.trace;
+
+import org.apache.skywalking.apm.agent.core.context.TracingContext;
+
+public abstract class StackBasedTracingSpan extends AbstractTracingSpan {
+    protected int stackDepth;
+    protected String peer;
+
+    protected StackBasedTracingSpan(int spanId, int parentSpanId, String operationName, TracingContext owner) {
+        super(spanId, parentSpanId, operationName, owner);
+        this.stackDepth = 0;
+        this.peer = null;
+    }
+
+    protected StackBasedTracingSpan(int spanId, int parentSpanId, String operationName, String peer,
+                                    TracingContext owner) {
+        super(spanId, parentSpanId, operationName, owner);
+        this.peer = peer;
+    }
+}