+++ ChangeLog.md
@@ -3,6 +3,9 @@
 1. All notable changes to this project will be documented in this file.
 2. Records in this file are not identical to the title of their Pull Requests. A detailed description is necessary for understanding what changes are and why they are made.
 
+## New Features
+- [Plugins] Enhance SkyWalking Agent and print endpoint info into /dev/null.
+
 ## v1.0.2 - 2023-01-03
 - [Plugins] Enhance Pinpoint Agent and print span into /dev/null.
 - [Plugins] Enhance Pinpoint Agent and print traceId into /dev/null.